Fix up commands
